feat(world): add deterministic infinite chunk streaming

+++ game/src/main/java/com/gaia/world/generation/BiomeProvider.java
@@ -1,5 +1,6 @@
 package com.gaia.world.generation;
 
 public interface BiomeProvider extends WorldGenerationStage {
-    BiomeSample sample(GenerationContext context, int worldX, int worldZ);
+    BiomeSample sample(
+            GenerationContext context, long worldX, long worldZ);
 }
