+++ game/src/main/java/com/gaia/world/generation/HeightProvider.java
@@ -3,7 +3,7 @@ package com.gaia.world.generation;
 public interface HeightProvider extends WorldGenerationStage {
     int sampleHeight(
             GenerationContext context,
-            int worldX,
-            int worldZ,
+            long worldX,
+            long worldZ,
             BiomeSample biome);
 }
